Starting refactor of 3d.ts into separate modules
