+++ src/3d/world/index.ts
@@ -1,0 +1,7 @@
+export const convertCoordsBlockToCoords = (coordsBlock: CoordsBlock) => {
+  if (!coordsBlock) return null;
+  if (!coordsBlock[0]) return null;
+  let coords = coordsBlock[0];
+  if (coords === null) return null;
+  return coords;
+};
