+++ src/blocks/lighting/moveLight.ts
@@ -28,6 +28,6 @@ export let moveLight = {
         let variable = javascriptGenerator.nameDB_.getName(block.getFieldValue("VAR"), "VARIABLE");
         if (coords === "") return "";
 
-        return `threeD.moveLight(${variable}, ${coords});`;
+        return `threeD.moveLight(${variable}, ${coords}, scene);`;
     }
 };
